s/HtmlSource/ClippedSource/g. Refact fragment parse. Triv testing refact.
Take a look at the way empty lines are being removed from HTML source. Not horrible, but I'm not sure if that should concern me or not.

See also the `default:` switch in the parameterized constructor. We're pretending any line with a colon that's not a known header format is the start of the HTML. I'm not sure that's smartest, though it's one way to allows us to process a fragment without an `<html` tag. (Not sure if that's necessarily a better way to parse, admittedly.) There are ways to be a little smarter (see if we have [a-zA-Z]+ before the `:` and keep going if it is, for instance. Hrm. I think I'll add that, but there are others) that I might want to consider.

TODO: Need to write some error case tests.